udong-backend/requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
Django==4.1.2
django-debug-toolbar==3.7.0
django-stubs==1.13.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-stubs==1.8.0
drf-yasg==1.21.4
drf-yasg-stubs==0.1.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
types-Markdown==3.4.2.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
